requirements.txt
openai>=1.0.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
redis>=5.0.0
